feat: hard-tier abilities — tests-first, 4 engine/parser fixes, compiler bug found
Wrote 15 tests FIRST from printed text for the 4 riskiest uncovered
abilities, then fixed everything needed to make them pass:

IDX 227 PL!HS-bp5-005-R 徒町小鈴 (cost-mirror + threshold heart):
- parser: _try_cost_set_from_reference — 「コストは選んだメンバーが元々持つ
  コストよりN低い/高い値に等しくなる。これによりこのカードのコストがM以上に
  なった場合…」 -> conditional_on_result[select, modify_cost(set_from_reference)]
  with location=activating_card threshold condition.
- engine: Operation::SetFromReference variant; execute_modify_cost branch
  resolving 選んだメンバー from selected_cards (NOT moved_cards — a cost
  discard would shadow it), delta = (ref original cost + offset) - target printed;
  COST_SELF comparison branch gated on location=activating_card so Rina's
  moved-cards-sum conditions (bp3-009) are not shadowed.
- COMPILER BUG (latent): enc_val wrote negative ints as raw <q with NO width
  byte; Rust read_int consumed the first payload byte as width -> whole tree
  desynced, rewrite pass IndexError'd. Now 0xFF + i32 two's complement both sides.

IDX 726 PL!SP-bp5-017-N 桜小路きな子 (常時 hand cost -2 while moved):
- state.rs HasMoved arm now honors group_names when scanning stage movement
  (non-Liella movers no longer grant the discount).

IDX 840 cl1-011-CL ド！ド！ド！ (choice options):
- choice executor no longer propagates parent group_names onto options that
  define their own (option A retrieves ANY member; option B stays 蓮ノ空-only).

IDX 461 bp6-003-R 松浦果南 (mill -> exact cost+2 rebirth): already worked;
pinned incl. no-match and exclude_self edges.

Diagnostics: MOD_COST_ENTRY/MOD_COST_REF/COST_SELF permanent logs.

Suite 2855/0+0ign. depth=none 40 -> 35.

+++ engine/src/ability/vm.rs
@@ -329,7 +329,8 @@ impl<'a> BcReader<'a> {
     }
 
     /// Read a TAG_I64 value with variable-width payload:
-    /// value ≤ 0xFD → 1 byte; 0xFE → +u16; 0xFF → +u32; negative → full i64.
+    /// value ≤ 0xFD → 1 byte; 0xFE → +u16; 0xFF → +i32 (two's complement,
+    /// signed so the compiler can encode negative ints such as cost_offset).
     fn read_int(&mut self) -> Option<i64> {
         let b = self.read_u8()?;
         if b <= 0xFD {
@@ -337,7 +338,7 @@ impl<'a> BcReader<'a> {
         } else if b == 0xFE {
             self.u16().map(|v| v as i64)
         } else if b == 0xFF {
-            self.read_u32().map(|v| v as i64)
+            self.read_u32().map(|v| v as i32 as i64)
         } else {
             self.i64()
         }
